feat: enable virtualization and add virt-manager, qemu, and OVMF packages
This commit enables virtualization by configuring the system.nix file to include the necessary packages and settings. The virtualization.libvirtd.enable option is set to true, and the qemu, virt-manager, and OVMF packages are added to the list of packages. This allows for the use of virtual machines on the system.